Implemented a first example of RTS Unit for the Green faction.
When manually generating the UGrid, DonkeyGun is able to move.
Removed warnings when a Unit does not have an Animation Controller.
Fixed movement field's initial reference within the UnitController.
Froze X, Y and Z rotations to prevent random non-RTS-like movement.
Changed the layer to Unit in order to enable collisions.
Changed the tag to Unit so that other scripts know how to interact.