Improved Pathfinding performance by preventing duplicate requests.
Used the amount of parallel A* threads as a limit of duplicates.
Prevented a special case of hazard under low FPS by extracting a
setter for UPath, that will now also reset starting position.
In future there may be a need for a more explicit synchronization.
Renamed a GetPath method that is not a getter, removed a Test class.