Implemented Unit production with multi-faction Resources support.
Units are now also properly created as Actors, not crashing the game.
The UnitProduced Event also contains a backreference to the Building.
BuildingCreated is now an independent Event from the creation start.
The same Actor is still valid after the finished Building gets placed.
The Unit that finished the Building construction is also available.
The Faction specific resources don't have to implement all of them,
because if the cost of some resource is 0, the index isn't check.
DonkeyGun is now a ranged Fighter.
Tweaked Units' values and doubled the default game speed.